refactor: consolidate AbstractMultifeatureDataclass implementation
- Unified `AbstractMultifeatureDataclass` and associated abstract dataclass
  utilities into dedicated modules to standardize usage and improve readability.
- move bag lvl timestamps collecting to TCT mf dataclass
- Deprecated and removed outdated functions, such as `collect_bag_lvl_timestamps`,
  to streamline the codebase and enhance maintainability.
- Updated imports across the project to align with the new module structure
  for consistent and clean organization.

Issue TCT-70, TCT-52
