+++ src/trajectory_container_tools/dataclasses/__init__.py
@@ -13,48 +13,51 @@ Usage:
     >>> imu_class = tct.dataclasses.ros_msgs.ros2_nested_dataclass.SensorMsgsImu
 """
 
-# Abstract base classes
-from trajectory_container_tools.dataclasses.core.abstract_trajectory_dataclass import (
+# Abstract base classes and base classes
+from trajectory_container_tools.dataclasses.core import (
     AbstractTrajectoryDataclass,
     AbstractMultifeatureDataclass,
-    AbstractNoTrajectoryDataclass
-)
-
-# Base implementations
-from trajectory_container_tools.dataclasses.core.base_trajectory_dataclass import (
+    AbstractNoTrajectoryDataclass,
     BaseTrajectoryDataclass,
     NestedBaseTrajectoryDataclass,
-    BaseNoTrajectoryDataclass
+    BaseNoTrajectoryDataclass,
 )
 
 # ROS2 dataclasses
 from trajectory_container_tools.dataclasses.ros_msgs.non_trajectory_dataclass import (
     Tf2MsgsTFMessage,
-    )
-from .ros_msgs.stamped_dataclass import AckermannMsgsAckermannDriveStamped, NavMsgsOdometry, \
-    Scan, VescMsgsVescImuStamped
+)
+from .ros_msgs.stamped_dataclass import (
+    AckermannMsgsAckermannDriveStamped,
+    NavMsgsOdometry,
+    Scan,
+    VescMsgsVescImuStamped,
+)
 from .ros_msgs.nested_dataclass import AckermannMsgsAckermannDrive, SensorMsgsImu
-from .ros_msgs.core_dataclass import NestedRosStampedDataclass, RosDataclass, \
-    RosStampedDataclass
+from .ros_msgs.core_dataclass import (
+    NestedRosStampedDataclass,
+    RosDataclass,
+    RosStampedDataclass,
+)
 
 from trajectory_container_tools.dataclasses.ros_msgs.primitive_dataclass import Header
 
-# Simulation environment dataclasses  
+# Simulation environment dataclasses
 from .f110_gym_trajectory_dataclass import (
     F110MotionDynamicDataclass,
     F110MotionDynamicDataclassNested,
     F110observations,
-    F110actions
+    F110actions,
 )
 from .math_gymnasium_trajectory_dataclass import (
     MathEnvTrajectoryDataclass,
     StateAxDataclass,
     TimeAxDataclass,
-    AxBaseDataclass
+    AxBaseDataclass,
 )
 from .erll_trajectory_dataclass import (
     TestTrajectoryDataclass,
-    TestMotionTrajectoryDataclass
+    TestMotionTrajectoryDataclass,
 )
 
 # DataFrame dataclasses
@@ -65,65 +68,55 @@ from .panda_dataframe_feature_dataclass import (
     CmdStandard,
     CmdSkidSteer,
     Velocity,
-    VelocitySkidSteer
+    VelocitySkidSteer,
 )
 
 # Primitive types
-from .primitive_dataclass import (
-    Point2D,
-    Vector2D,
-    Pose2D,
-    Velocity2D
-)
+from .primitive_dataclass import Point2D, Vector2D, Pose2D, Velocity2D
 
 __all__ = [
     # Abstract classes
-    'AbstractTrajectoryDataclass',
-    'AbstractMultifeatureDataclass',
-    'AbstractNoTrajectoryDataclass',
-    
+    "AbstractTrajectoryDataclass",
+    "AbstractMultifeatureDataclass",
+    "AbstractNoTrajectoryDataclass",
     # Base classes
-    'BaseTrajectoryDataclass', 
-    'NestedBaseTrajectoryDataclass',
-    'BaseNoTrajectoryDataclass',
-    
+    "BaseTrajectoryDataclass",
+    "NestedBaseTrajectoryDataclass",
+    "BaseNoTrajectoryDataclass",
     # ROS2 dataclasses
-    'NavMsgsOdometry',
-    'SensorMsgsImu', 
-    'AckermannMsgsAckermannDriveStamped',
-    'AckermannMsgsAckermannDrive',
-    'Tf2MsgsTFMessage',
-    'VescMsgsVescImuStamped',
-    'Scan',
-    'RosStampedDataclass',
-    'RosDataclass',
-    'NestedRosStampedDataclass',
-    'Header',
-    
+    "NavMsgsOdometry",
+    "SensorMsgsImu",
+    "AckermannMsgsAckermannDriveStamped",
+    "AckermannMsgsAckermannDrive",
+    "Tf2MsgsTFMessage",
+    "VescMsgsVescImuStamped",
+    "Scan",
+    "RosStampedDataclass",
+    "RosDataclass",
+    "NestedRosStampedDataclass",
+    "Header",
     # Simulation environments
-    'F110MotionDynamicDataclass',
-    'F110MotionDynamicDataclassNested',
-    'F110observations',
-    'F110actions',
-    'MathEnvTrajectoryDataclass',
-    'StateAxDataclass',
-    'TimeAxDataclass',
-    'AxBaseDataclass',
-    'TestTrajectoryDataclass',
-    'TestMotionTrajectoryDataclass',
-    
+    "F110MotionDynamicDataclass",
+    "F110MotionDynamicDataclassNested",
+    "F110observations",
+    "F110actions",
+    "MathEnvTrajectoryDataclass",
+    "StateAxDataclass",
+    "TimeAxDataclass",
+    "AxBaseDataclass",
+    "TestTrajectoryDataclass",
+    "TestMotionTrajectoryDataclass",
     # DataFrame
-    'BaseDataframeFeatureDataclass',
-    'NestedBaseDataframeFeatureDataclass',
-    'StatePose2D',
-    'CmdStandard',
-    'CmdSkidSteer',
-    'Velocity',
-    'VelocitySkidSteer',
-    
+    "BaseDataframeFeatureDataclass",
+    "NestedBaseDataframeFeatureDataclass",
+    "StatePose2D",
+    "CmdStandard",
+    "CmdSkidSteer",
+    "Velocity",
+    "VelocitySkidSteer",
     # Primitives
-    'Point2D',
-    'Vector2D',
-    'Pose2D',
-    'Velocity2D',
+    "Point2D",
+    "Vector2D",
+    "Pose2D",
+    "Velocity2D",
 ]
