+++ src/trajectory_container_tools/dataclasses/core/__init__.py
@@ -1,1 +1,9 @@
 # coding=utf-8
+from .abstract_multi_trajectory_dataclass import AbstractMultifeatureDataclass
+from .abstract_no_trajectory_dataclass import AbstractNoTrajectoryDataclass
+from .abstract_trajectory_dataclass import AbstractTrajectoryDataclass
+from .base_trajectory_dataclass import (
+    BaseTrajectoryDataclass,
+    BaseNoTrajectoryDataclass,
+    NestedBaseTrajectoryDataclass,
+)
