+++ src/trajectory_container_tools/dataclasses/core/abstract_trajectory_dataclass.py
@@ -1,191 +1,17 @@
 # coding=utf-8
-import abc
-import datetime
 from copy import deepcopy
-from dataclasses import dataclass, field, fields
+from dataclasses import dataclass, field
 import numpy as np
-from typing import Any, List, Optional, Tuple, Union
+from typing import List, Union
 
-import trajectory_container_tools as tct
+from trajectory_container_tools.dataclasses.core.abstract_trajectory_dataclass_common import (
+    AbstractTrajectoryDataclassCommon,
+)
 from trajectory_container_tools.temporal import Timestamps
 from trajectory_container_tools.temporal import validate_timestep_indices
 from trajectory_container_tools.utils.general import (
-    check_typing_list_and_extract_list_type,
     extract_class_name_from_instance,
-    check_typing_union_and_extract_first_union_type,
 )
-
-
-@dataclass()
-class AbstractTrajectoryDataclassCommon(abc.ABC):
-    """
-    Provides an abstract base dataclass for dynamically interacting with and managing object
-    attributes, including dynamic fields creation and nested attribute retrieval.
-
-    This class serves as a foundation for implementing objects that must support
-    dynamically added fields, as well as hierarchical retrieval of nested attributes. It is
-    an abstract base class (ABC) and requires implementation of the `__post_init__` method
-    in subclasses.
-
-    """
-
-    @classmethod
-    def _dataclass_internal_field(cls) -> List[str]:
-        """
-        List of field marked as internal. Those are field name that will be omited by
-        `get_dimension_names` class method.
-
-        This method is intended to return a predefined list of attribute names that are
-        specific to the internal logic of a data class. These fields often represent
-        key information required for specialized operations or manipulations within
-        the class. The method should be used internally and not be exposed for
-        general use.
-
-        :return: A list containing the names of internal fields used in the data class.
-        """
-        return []
-
-    def get_dynamic_field(self, feature_name: str) -> Any:
-        """Retrieves the value of a dynamicaly declared attribute from the object.
-
-        :param feature_name: The name of the attribute to retrieve.
-        :return: The value of the requested attribute.
-        """
-        return self.__getattribute__(feature_name)
-
-    def set_dynamic_field(self, feature_name: str, value: Any) -> None:
-        """Updates or creates a dynamic attribute on an object.
-
-        :param feature_name: The name of the attribute to update or create.
-        :param value: The value to assign to the attribute.
-        :return: None
-        """
-        self.__setattr__(feature_name, value)
-        return None
-
-    def fetch_nested_attribute(self, nested_attribute_path: str) -> Any:
-        """Retrieves a nested attribute from an object based on a dot-separated string.
-
-        This function allows accessing nested attributes of an object dynamically, based on a
-        string representation of the attribute's hierarchical structure.
-        It takes a dot-separated attribute name, traverses the object's nested levels
-        sequentially, and retrieves the final attribute
-        e.g., "topic_odom.pose.pose.position.x" would sequentialy crawl into nested container
-        "topic_odom" -> "pose" -> "pose" -> "position" -> "x".
-
-        Example:
-
-            >>> position_x_value = self.fetch_nested_attribute("topic_odom.pose.pose.position.x")
-
-        :param nested_attribute_path: A dot-separated string representing the hierarchical
-          structure of the attribute to retrieve.
-        :return: The value of the resolved nested attribute.
-        """
-        return _fetch_nested_attribute(self, nested_attribute_path)
-
-    @abc.abstractmethod
-    def __post_init__(self):
-        pass
-
-    @classmethod
-    def get_dimension_type(cls, dimension_name: str) -> Tuple[type[Any], bool]:
-        """
-        Get the type and whether it is a list for a given dimension name.
-
-        The method inspects the class fields, determining the type of the field associated with the provided dimension name. It identifies whether the type is a list or a union, returning the underlying type if applicable.
-
-        :param dimension_name: The name of the dimension to query.
-        :return: A tuple containing the type of the dimension and a boolean indicating if the dimension is a list.
-        """
-        container_properties = fields(cls)
-        dimension_type = None
-        is_list_of_type = False
-        for each_field in container_properties:
-            if each_field.name is dimension_name:
-                is_list_of_type, dimension_type = (
-                    check_typing_list_and_extract_list_type(each_field.type)
-                )
-                if not is_list_of_type:
-                    is_union, dimension_type = (
-                        check_typing_union_and_extract_first_union_type(each_field.type)
-                    )
-        return dimension_type, is_list_of_type
-
-    @classmethod
-    def get_dimension_names(cls) -> Tuple[str, ...]:
-        """
-        Provides the dimension names of the data class fields that are exposed to user.
-
-        This method retrieves the field names from the data class that are not marked
-        as internal fields. It returns these names as a tuple of strings, making it
-        useful for querying the explicitly defined data attributes of the class.
-
-        :return: A tuple containing the names of non-internal fields defined in the class.
-        """
-        container_properties = fields(cls)
-        field_name = []
-        for each_field in container_properties:
-            if each_field.name not in cls._dataclass_internal_field():
-                field_name.append(each_field.name)
-        return tuple(field_name)
-
-
-@dataclass()
-class AbstractNoTrajectoryDataclass(AbstractTrajectoryDataclassCommon):
-    """
-    Representation of an abstract data structure for entities without trajectory data at top-level
-    but might in nested ones e.g., a container that contains many trajectory dataclass of different
-    trajectory lenghts.
-    """
-    feature_name: str
-
-    @classmethod
-    def _dataclass_internal_field(cls) -> List[str]:
-        return [
-            "feature_name",
-        ]
-
-    def __str__(self):
-        """User representation. Dynamically handle property added at run time"""
-        t_sp = " " * 10
-        m_sp = " " * 3
-        item_space = " " * 3
-        dataclass_name = extract_class_name_from_instance(self)
-        repr_str = f"\n{t_sp}{dataclass_name}(\n"
-        v: Union[np.ndarray, AbstractTrajectoryDataclass, str, int, float]
-        m_sp += t_sp
-
-        v = self.__dict__.get("feature_name")
-        if v is not None:
-            repr_str += f"{m_sp}feature_name: {v}\n"
-
-        for k, v in self.__dict__.items():
-            if k in [
-                "feature_name",
-            ]:
-                pass
-            elif isinstance(v, list):
-                indent_v = []
-                for each in v:
-                    for each_line in str(each).splitlines():
-                        indent_v.append(f"{t_sp}{each_line}")
-                    indent_v[-1] = f"{indent_v[-1]},"
-                indent_v = "\n".join(indent_v)
-                repr_str += f"{m_sp}{item_space}{k}: [" f"{indent_v}" f"\n{m_sp}]\n"
-
-            elif isinstance(v, AbstractTrajectoryDataclass):
-                indent_v = []
-                for each_line in str(v).splitlines():
-                    indent_v.append(f"{t_sp}{each_line}\n")
-                indent_v = "".join(indent_v)
-                repr_str += f"{m_sp}{item_space}{k}:{indent_v}"
-            else:
-                repr_str += f"{m_sp}{item_space}{k}: ({extract_class_name_from_instance(v)}) {v}\n"
-        repr_str += f"{t_sp})"
-        return repr_str
-
-    def __post_init__(self):
-        pass
 
 
 @dataclass()
@@ -570,69 +396,6 @@ class AbstractTrajectoryDataclass(AbstractTrajectoryDataclassCommon):
         # noinspection PyAttributeOutsideInit
         self._transposed = not self._transposed
         return self
-
-
-@dataclass
-class AbstractMultifeatureDataclass(AbstractTrajectoryDataclassCommon):
-    dataset_info: str
-    aggregated_date: datetime.datetime = field(init=False)
-    bag_timestamps: Optional[Timestamps] = field(default=None, kw_only=True)
-
-    @classmethod
-    def _dataclass_internal_field(cls) -> List[str]:
-        return []
-
-    def __post_init__(self):
-        self.aggregated_date = datetime.datetime.now()
-
-    def __str__(self):
-        """User representation. Handle dynamical property added at run time"""
-        t_sp = " " * 0
-        m_sp = " " * 3
-        repr_str = f"\n{t_sp}Multifeature(\n"
-        m_sp += t_sp
-        for k, v in self.__dict__.items():
-            if k == "dataset_info":
-                repr_str += f"{m_sp}dataset_info: {v}\n"
-                repr_str += f"{m_sp}aggregated_date: {self.aggregated_date}\n"
-            elif k in ["aggregated_date"]:
-                pass
-            elif k in ["bag_timestamps"] and self.bag_timestamps is None:
-                pass
-            elif isinstance(v, (np.ndarray, Timestamps)):
-                if isinstance(v, Timestamps):
-                    range_str = (
-                        f"range(nanosec) {np.min(v.stamps)} ⟶ {np.max(v.stamps)}"
-                    )
-                else:
-                    range_str = f"range {np.min(v)} ⟶ {np.max(v)}"
-                repr_str += (
-                    f"{m_sp}{k}: ({extract_class_name_from_instance(v)}) "
-                    f"shape {v.shape} {range_str}\n"
-                )
-            else:
-                repr_str += f"{m_sp}{k}: {str(v)}\n"
-        repr_str += f"{m_sp})"
-        return repr_str
-
-    @property
-    def summary(self) -> None:
-        # (NICE TO HAVE) ToDo: TCT-68 feat: deprecate AbstractMultifeatureDataclass summary property
-        print(self)
-        return None
-
-    @property
-    def topic_key_list(self):
-        return [
-            topic.name for topic in fields(self) if str(topic.name).startswith("topic_")
-        ]
-
-
-def _fetch_nested_attribute(self_, nested_attribute_list: str) -> Any:
-    nested_attribute = self_
-    for each in nested_attribute_list.split("."):
-        nested_attribute = nested_attribute.get_dynamic_field(each)
-    return nested_attribute
 
 
 def _timesteps_indices_vs_index_len_check(
