+++ src/trajectory_container_tools/extractor/dataframe_to_tct.py
@@ -7,9 +7,7 @@ from typing import Dict, Tuple, Union
 from dataclasses import make_dataclass
 
 from trajectory_container_tools.typing import MultifeatureTrajectoryDataclass
-from trajectory_container_tools.dataclasses.core.abstract_trajectory_dataclass import (
-    AbstractMultifeatureDataclass,
-)
+from trajectory_container_tools import AbstractMultifeatureDataclass
 from trajectory_container_tools.dataclasses.panda_dataframe_feature_dataclass import (
     BaseDataframeFeatureDataclass,
 )
