+++ tests/tests_utils/test_rosbag_eda.py
@@ -5,6 +5,7 @@ import shutil
 import pytest
 
 import trajectory_container_tools as tct
+import trajectory_container_tools.dataclasses.core.abstract_multi_trajectory_dataclass
 from trajectory_container_tools.utils.ros2_utils.rosbag_eda.rosbag_timestamp_eda import (
     run_rosbag_timestamp_eda,
 )
@@ -79,7 +80,8 @@ def test_run_rosbag_timestamp_eda_full_bag(
         show_plot=False,
     )
     print(tc)
-    assert isinstance(tc, tct.AbstractMultifeatureDataclass)
+    assert isinstance(tc,
+                      trajectory_container_tools.dataclasses.core.abstract_multi_trajectory_dataclass.AbstractMultifeatureDataclass)
 
     assert tc.topic_odom.trajectory_len == 864
     assert tc.topic_tf.transforms[0].trajectory_len == 866
@@ -102,7 +104,8 @@ def test_run_rosbag_timestamp_eda_window(
         show_plot=False,
     )
     print(tc)
-    assert isinstance(tc, tct.AbstractMultifeatureDataclass)
+    assert isinstance(tc,
+                      trajectory_container_tools.dataclasses.core.abstract_multi_trajectory_dataclass.AbstractMultifeatureDataclass)
 
     assert tc.topic_odom.trajectory_len == 364
     assert tc.topic_tf.transforms[0].trajectory_len == 364
